# specs/001-todo-fullstack-app/contracts/api.yaml
openapi: 3.1.0
info:
  title: Todo Full-Stack API
  description: |
    RESTful API for the Todo Full-Stack Web Application.
    All /tasks endpoints require JWT authentication via Bearer token.
  version: 1.0.0
  contact:
    name: Hackathon Phase II

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.todo-app.example.com
    description: Production

security:
  - bearerAuth: []

tags:
  - name: Health
    description: Health check endpoints
  - name: Tasks
    description: Task management endpoints (authenticated)

paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns server health status. No authentication required.
      security: []
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/tasks:
    get:
      tags:
        - Tasks
      summary: List user's tasks
      description: Returns all tasks owned by the authenticated user.
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Tasks
      summary: Create a task
      description: Creates a new task for the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/tasks/{task_id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'

    get:
      tags:
        - Tasks
      summary: Get a task
      description: Returns a single task by ID. Only returns tasks owned by authenticated user.
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Tasks
      summary: Update a task
      description: Updates a task's title, description, and/or completion status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: Permanently deletes a task. Only deletes tasks owned by authenticated user.
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/tasks/{task_id}/toggle:
    parameters:
      - $ref: '#/components/parameters/TaskId'

    patch:
      tags:
        - Tasks
      summary: Toggle task completion
      description: Toggles the is_completed status of a task.
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      description: UUID of the task
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"

  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: "ok"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title (required)
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          description: Optional task description
          example: "Milk, eggs, bread"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated task title
          example: "Buy groceries and snacks"
        description:
          type: string
          nullable: true
          description: Updated task description
          example: "Milk, eggs, bread, chips"
        is_completed:
          type: boolean
          description: Updated completion status
          example: true

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          description: Owner's user ID
          example: "user_abc123"
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          description: Task description
          example: "Milk, eggs, bread"
        is_completed:
          type: boolean
          description: Whether task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-01-30T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-01-30T10:00:00Z"

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Task not found"

    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of the error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type

  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Not authenticated"

    NotFound:
      description: Resource not found or not owned by user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Task not found"

    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            detail:
              - loc: ["body", "title"]
                msg: "field required"
                type: "value_error.missing"